Add small comments, put NewSlotForm and corresponding tests in it's own folder, create test ids and errorMessages for SlotForm as constants
