+++ eisbuk-admin/src/components/temp/NewSlotForm/__testData__/testIds.ts
@@ -1,0 +1,11 @@
+/**
+ * @Ivan : For some reason it makes sense to me to write test ids as "lowercase-hyphen-divided-string". I guess thats because we write test ids as `data-testid`
+ * @NOTE Again, a personal preference
+ */
+export const __startTimeInputId__ = "start-time-input";
+
+export const __startTimeErrorId__ = "start-time-error";
+
+export const __endTimeInputId__ = "end-time-input";
+
+export const __endTimeErrorId__ = "end-time-error";
