+++ eisbuk-admin/src/lib/errorMessages.ts
@@ -1,0 +1,3 @@
+// #region SlotForm
+export const __noStartTimeError = "Start Time is required";
+// #endregion SlotForm
